refactor: restructure project files, so that it looks like Python module
